update e2e tests
