+++ test/testkit/metrics/runtime/default.go
@@ -2,4 +2,13 @@ package runtime
 
 import "slices"
 
-var DefaultMetricsNames = slices.Concat(ContainerMetricsNames, PodMetricsNames)
+var DefaultMetricsNames = slices.Concat(
+	ContainerMetricsNames,
+	PodMetricsNames,
+	VolumeMetricsNames,
+	NodeMetricsNames,
+	DeploymentMetricsNames,
+	DaemonSetMetricsNames,
+	StatefulSetMetricsNames,
+	JobsMetricsNames,
+)
